requirements.txt
PyGithub>=2.9.1
python-dotenv>=1.1.0
requests>=2.32.3
Flask>=3.1.1
pytest>=8.4.2
pytest-mock>=3.15.0
requests-mock>=1.12.1
